% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-init.R
\docType{package}
\name{accept-package}
\alias{accept}
\alias{accept-package}
\title{accept: The Acute COPD Exacerbation Prediction Tool (ACCEPT)}
\description{
Allows clinicians to predict the rate and severity of future acute exacerbation in Chronic Obstructive Pulmonary Disease (COPD) patients, based on the clinical prediction models published in Adibi et al. (2020) \doi{10.1016/S2213-2600(19)30397-2} and Safari et al. (2022) \doi{10.1016/j.eclinm.2022.101574}.
}
\author{
\strong{Maintainer}: Amin Adibi \email{adibi@alumni.ubc.ca}

Authors:
\itemize{
  \item Mohsen Sadatsafavi \email{mohsen.sadatsafavi@ubc.ca} [copyright holder]
  \item Abdollah Safari \email{asafari@mail.ubc.ca}
  \item Ainsleigh Hill \email{ainsleigh.hill@alumni.ubc.ca}
}

}
\keyword{internal}
