% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{estimateACCEPT}
\alias{estimateACCEPT}
\title{Predicts COPD exacerbations within the next year}
\usage{
estimateACCEPT(patientData, random_sampling_N = 10000)
}
\arguments{
\item{patientData}{patient data matrix. Can have one or many patients in it}

\item{random_sampling_N}{number of random sampling. Default is 1000.}
}
\value{
patientData with prediction
}
\description{
Predicts COPD exacerbations within the next year
}
\examples{
results <- predictACCEPT(samplePatients)
}
