% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fev1_predictor}
\alias{fev1_predictor}
\title{predict fev1 for the next 15 years}
\usage{
fev1_predictor(data)
}
\arguments{
\item{data}{a data frame with patients characteristics including height and weight}
}
\value{
a data frame contains average annual FEV1, its variance, and a 95% CI.
}
\description{
predict fev1 for the next 15 years
}
\examples{
data_temp <- cbind(samplePatients,
                   height = 1.7,
                   weight = 70)
results <- fev1_predictor(data = data_temp[1, ])
}
