% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{FlexACCEPT}
\alias{FlexACCEPT}
\title{A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.}
\usage{
FlexACCEPT(data, refData, num_imput = 10, max_iter = 5, seed = NA)
}
\arguments{
\item{data}{new patient data with missing values to be imputed before prediction with the same format as accept samplePatients.}

\item{refData}{a referenced data with the same format as accept samplePatients that the imputation will be based on.}

\item{num_imput}{Number of multiple imputations. The default is 10.}

\item{max_iter}{A scalar giving the number of iterations. The default is 5.}

\item{seed}{An integer that is used as argument by the set.seed() for offsetting the random number generator. Default is to leave the random number generator alone.}
}
\value{
patientData with prediction.
}
\description{
A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.
}
\examples{
results <- FlexACCEPT(data = samplePatients, refData = samplePatients)
}
