% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{FlexCCEPT}
\alias{FlexCCEPT}
\title{A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.}
\usage{
FlexCCEPT(data)
}
\arguments{
\item{data}{new patient data with missing values to be imputed before prediction with the same format as accept samplePatients.}
}
\value{
patientData with prediction.
}
\description{
A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.
}
\examples{
results <- FlexCCEPT(data = samplePatients)
}
