% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{Creates heatmap of number of exacerbations}
\usage{
plotHeatMap(patientResults, n = 10)
}
\arguments{
\item{patientResults}{patient results vector, produced by predictAccept.}

\item{n}{how many exacerbations to consider}

\item{shortened}{boolean}
}
\value{
a matrix of probabilities with the number of exacerbations as rows and number of severe exacerbations as columns
}
\description{
Creates heatmap of number of exacerbations
}
\examples{
results <- predictACCEPT(samplePatients[1,])
plotHeatMap(results)
}
