% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-compatibility.R
\name{predict.accept_model}
\alias{predict.accept_model}
\title{Model predictions}
\usage{
\method{predict}{accept_model}(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{new data for which the prediction is needed}
}
\value{
prediction results from ACCEPT model
}
\description{
Model predictions
}
