% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sample_enrichment}
\alias{sample_enrichment}
\title{sample enrichment function to find optimal sample size or threshold relative risk for patient recruitment}
\usage{
sample_enrichment(
  test_alpha = 0.05,
  test_beta = 0.2,
  trt_prop = 0.5,
  trt_effect = log(1/0.85),
  event_rate,
  n = NULL,
  sample_size_range = FALSE,
  thresholds = NULL,
  plot = FALSE
)
}
\arguments{
\item{test_alpha}{type I error of the test (default = 0.05)}

\item{test_beta}{type II error of test (default = 0.2)}

\item{trt_prop}{proportion of patients in treatment arm (default = 0.5)}

\item{trt_effect}{treatment effect to be aimed from the data analysis}

\item{event_rate}{event rate of the outcome in the target population}

\item{n}{sample size  (default = NULL)}

\item{sample_size_range}{logical, whether a range of sample size being returned corresponding to the deciles of the predicted risk}

\item{thresholds}{threshold values at which sample size needs to be estimated}

\item{plot}{plot estimated sample size versus relative risk thresholds and recruitment efficiency}
}
\value{
either a relative risk threshold value corresponds to a given sample size or return a range of sample sizes corresponding to different threshold values of relative risk
}
\description{
sample enrichment function to find optimal sample size or threshold relative risk for patient recruitment
}
\examples{
## finding optimal relative risk threshold value corresponding to a given sample size
sample_enrichment(event_rate = 0.4, n = 400)
## estimating a range of sample sizes corresponding to different values of relative risk
sample_enrichment(event_rate = 0.4, sample_size_range = TRUE)
}
