% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{txRecom}
\alias{txRecom}
\title{CTS Treatment Guideline based on Mild, Moderate, or Severe Categories}
\usage{
txRecom(patient)
}
\arguments{
\item{patient}{patient information vector that includes predictions. must include either SGRQ or CAT score}
}
\value{
a message that specifies treatment recommendation based on the threshold specified and CTS guidelines
}
\description{
CTS Treatment Guideline based on Mild, Moderate, or Severe Categories
}
\examples{
patient <- accept(samplePatients[1,])
patientPredictions <- accept2(patient)
txRecom (patientPredictions)
}
