% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-compatibility.R
\name{vetiver_create_description.accept_model}
\alias{vetiver_create_description.accept_model}
\title{Create a vetiver model object description}
\usage{
\method{vetiver_create_description}{accept_model}(model)
}
\arguments{
\item{model}{model name}
}
\value{
the description of the model
}
\description{
Create a vetiver model object description
}
