% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-compatibility.R
\name{vetiver_create_meta.accept_model}
\alias{vetiver_create_meta.accept_model}
\title{Metadata constructors for `vetiver_model()` object}
\usage{
\method{vetiver_create_meta}{accept_model}(model, metadata)
}
\arguments{
\item{model}{a trained model object}

\item{metadata}{list containing additional metadata to store with the pin}
}
\value{
The vetiver_create_meta function returns a `vetiver_meta()` list.
}
\description{
Useful for supporting accept_model class in vetiver
}
