% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-compatibility.R
\name{vetiver_ptype.accept_model}
\alias{vetiver_ptype.accept_model}
\title{Create a vetiver input data prototype}
\usage{
\method{vetiver_ptype}{accept_model}(model, ...)
}
\arguments{
\item{model}{a vetiver model object}

\item{...}{any other inputs}
}
\value{
A `vetiver_ptype` method returns a zero-row dataframe, and
`vetiver_create_ptype()` returns either such a zero-row dataframe, `NULL`,
or the dataframe passed to `save_prototype`.
}
\description{
Create a vetiver input data prototype
}
