% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{accept}
\alias{accept}
\title{A flexible version of ACCEPT 2.0 model, which imputes predictors using MICE approach.}
\usage{
accept(
  newdata,
  format = "tibble",
  version = "accept3",
  prediction_interval = FALSE,
  return_predictors = FALSE,
  country = NULL,
  ...
)
}
\arguments{
\item{newdata}{new patient data with missing values to be imputed before prediction with the same format as accept samplePatients.}

\item{format}{default is "tibble". Can also be set to "json".}

\item{version}{indicates which version of ACCEPT needs to be called. Options include "accept1", "accept2", "accept3" (default).}

\item{prediction_interval}{default is FALSE. If set to TRUE, returns prediction intervals of the predictions.}

\item{return_predictors}{default is FALSE. IF set to TRUE, returns the predictors along with prediction results.}

\item{country}{Required for accept3 version. Three-letter ISO country code (e.g., "CAN", "USA", "GBR"). Supported countries: ARG, AUS, BRA, CAN, COL, DEU, DNK, ESP, FRA, GBR, ITA, JPN, KOR, MEX, NLD, NOR, SWE, USA. For unsupported countries, provide obs_modsev_risk in the data.}

\item{...}{for other versions of accept.}
}
\value{
patientData with prediction.
}
\description{
A flexible version of ACCEPT 2.0 model, which imputes predictors using MICE approach.
}
\examples{
results <- accept(newdata = samplePatients, country = "CAN")
results_us <- accept(newdata = samplePatients, country = "USA")
}
