% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{accept}
\alias{accept}
\title{A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.}
\usage{
accept(data, version = "flexccept", prediction_interval = FALSE, ...)
}
\arguments{
\item{data}{new patient data with missing values to be imputed before prediction with the same format as accept samplePatients.}

\item{version}{indicates which version of ACCEPT needs to be called.}

\item{prediction_interval}{default is FALSE If set to true, returns prediction intervals of the predictions.}

\item{...}{for other versions of accept.}
}
\value{
patientData with prediction.
}
\description{
A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.
}
\examples{
results <- accept(data = samplePatients)
}
