% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{accept}
\alias{accept}
\title{Predicts COPD exacerbation rate by severity level}
\usage{
accept(
  patientData,
  random_sampling_N = 100,
  lastYrExacCol = "LastYrExacCount",
  lastYrSevExacCol = "LastYrSevExacCount",
  calculate_CIs = FALSE,
  ...
)
}
\arguments{
\item{patientData}{patient data matrix. Can have one or many patients in it}

\item{random_sampling_N}{number of random sampling. Default is 100.}

\item{lastYrExacCol}{the column specifying last year all exacerbation count}

\item{lastYrSevExacCol}{the column specifying last year severe exacerbation count}

\item{calculate_CIs}{whether to calculate confidence interval of the mean}
}
\value{
patientData with prediction
}
\description{
Predicts COPD exacerbation rate by severity level
}
\examples{
results <- accept(samplePatients)
}
