% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotExacerbations}
\alias{plotExacerbations}
\title{Creates bar graph}
\usage{
plotExacerbations(patientResults, type = "rate", interval = "CI",
  colors = c("#007bff", "rgb(204,204,204)"))
}
\arguments{
\item{patientResults}{patient results vector, produced by predictAccept.}

\item{type}{string: either "rate" or "probability"}

\item{interval}{string: either "CI" or "PI"
CI = confidence interval
PI = predicted interval}

\item{colors}{vector: vector of color strings, rbg or hexadecimal, for colors on bar graph}
}
\value{
a bar graph
}
\description{
Creates bar graph
}
\examples{
results <- predictACCEPT(samplePatients[1,])
plotExacerbations(results)
}
