% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotExacerbations}
\alias{plotExacerbations}
\title{Creates bar graph comparing no treatment with azithromycin treatment}
\usage{
plotExacerbations(
  patientResults,
  type = "rate",
  interval = "CI",
  colors = c("#007bff", "rgb(204,204,204)")
)
}
\arguments{
\item{patientResults}{patient results vector, produced by predictAccept.}

\item{type}{string: either "probability" or "rate"}

\item{interval}{string: either "CI" or "PI"
PI = Predicted Interval
CI = Confidence Interval}

\item{colors}{vector: a vector of colors to be used in the graph
must be length 2
can use hexadecimal, rgb, or R color codes}
}
\value{
a bar graph
}
\description{
Creates bar graph comparing no treatment with azithromycin treatment
}
\examples{
results <- predictACCEPT(samplePatients[1,], random_distribution_iteration = 5000)
plotExacerbations(results)
}
