% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{Creates heatmap of number of exacerbations}
\usage{
plotHeatMap(patientResults, n = 10, shortened = TRUE)
}
\arguments{
\item{patientResults}{patient results vector, produced by accept.}

\item{n}{how many exacerbations to consider}

\item{shortened}{boolean}
}
\value{
a heatmap
}
\description{
Creates heatmap of number of exacerbations
}
\examples{
\dontrun{
results <- accept1(samplePatients[1,])
plotHeatMap(results)
}
}
