% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictACCEPT}
\alias{predictACCEPT}
\title{Initializes a model. Allocates memory to the C engine.}
\usage{
predictACCEPT(patientData, random_sampling_N = 1000)
}
\arguments{
\item{patientData}{patient data matrix. Can have one or many patients in it}

\item{random_sampling_N}{number of random sampling. Default is 1000.}
}
\value{
patientData with prediction
}
\description{
Initializes a model. Allocates memory to the C engine.
}
