% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictACCEPT}
\alias{predictACCEPT}
\title{Predicts COPD exacerbation rate by severity level}
\usage{
predictACCEPT(
  patientData,
  random_sampling_N = 1000,
  random_distribution_iteration = 20000,
  calculate_CIs = TRUE
)
}
\arguments{
\item{patientData}{patient data matrix. Can have one or many patients in it}

\item{random_sampling_N}{number of random sampling. Default is 1000.}

\item{random_distribution_iteration}{default is 2*10^4}

\item{calculate_CIs}{whether to calculate confidence interval of the mean}
}
\value{
patientData with prediction
}
\description{
Predicts COPD exacerbation rate by severity level
}
\examples{
results <- predictACCEPT(samplePatients, random_distribution_iteration = 5000)
}
