% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictCountProb}
\alias{predictCountProb}
\title{Predicts probability of observing a certain number of exacerbations}
\usage{
predictCountProb(patientResults, n = 10, shortened = TRUE)
}
\arguments{
\item{patientResults}{patient results matrix, produced by predictAccept.}

\item{n}{how many exacerbations to consider}

\item{shortened}{TODO}
}
\value{
a matrix of probabilities with the number of exacerbations as rows and number of severe exacerbations as columns
}
\description{
Predicts probability of observing a certain number of exacerbations
}
\examples{
results <- predictACCEPT(samplePatients[1,])
predictCountProb (results)
}
